ingest_rfm.py: Don't overspam events with whole removed and added list

+++ ingest_rfm.py
@@ -6,12 +6,18 @@ from generate_events import PersonEventGenerator
 engine = create_engine('postgresql://db@localhost:5432/postgres')
 
 with engine.begin() as connection:
-    connection.execute(text('TRUNCATE TABLE out.rfm_person'))
-    connection.execute(text('INSERT INTO out.rfm_person SELECT * from db.rfm_person'))
-    connection.execute(text('TRUNCATE TABLE db.rfm_person'))
-    connection.execute(text('INSERT INTO db.rfm_person SELECT * from scraper.rfm_person'))
-    connection.execute(text('TRUNCATE TABLE scraper.rfm_person'))
-    connection.commit()
+    scraper_result = connection.execute(text('SELECT COUNT(*) FROM scraper.rfm_person')).scalar()
+    db_result = connection.execute(text('SELECT COUNT(*) FROM db.rfm_person')).scalar()
 
-    generator = PersonEventGenerator(sql_engine=engine, mongo_client=MongoClient())
-    generator.generate()
+    if scraper_result > 0:  # Prevent every person truncation for unsuccessful scraping
+        connection.execute(text('TRUNCATE TABLE out.rfm_person'))
+        connection.execute(text('INSERT INTO out.rfm_person SELECT * from db.rfm_person'))
+        connection.execute(text('TRUNCATE TABLE db.rfm_person'))
+        connection.execute(text('INSERT INTO db.rfm_person SELECT * from scraper.rfm_person'))
+        connection.execute(text('TRUNCATE TABLE scraper.rfm_person'))
+
+        if db_result > 0:  # Ignore every person 'added' events generation for first start
+            generator = PersonEventGenerator(connection=connection, mongo_client=MongoClient())
+            generator.generate()
+
+    connection.commit()
